izmena tipa id atributa

+++ ISA_AMA_projekat/src/main/java/com/example/ISA_AMA_projekat/model/Korisnik.java
@@ -5,10 +5,8 @@ import static javax.persistence.FetchType.LAZY;
 
 import java.io.Serializable;
 import java.sql.Timestamp;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
@@ -30,7 +28,6 @@ import org.joda.time.DateTime;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
@@ -43,7 +40,7 @@ public class Korisnik implements Serializable, UserDetails
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@Column(unique = true, nullable = false)
 	private String email;
@@ -131,12 +128,12 @@ public class Korisnik implements Serializable, UserDetails
 
 	//GET & SET:
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
