+++ ISA_AMA_projekat/src/main/java/com/example/ISA_AMA_projekat/repository/FilijalaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ISA_AMA_projekat.model.Filijala;
 
-public interface FilijalaRepository extends JpaRepository<Filijala, Long>{
+public interface FilijalaRepository extends JpaRepository<Filijala, Integer>{
 
 }
