+++ ISA_AMA_projekat/src/main/java/com/example/ISA_AMA_projekat/repository/RezervacijaVozilaRepository.java
@@ -7,10 +7,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.ISA_AMA_projekat.model.RezervacijaVozila;
 
-public interface RezervacijaVozilaRepository extends JpaRepository<RezervacijaVozila, Long>{
+public interface RezervacijaVozilaRepository extends JpaRepository<RezervacijaVozila, Integer>{
 
 	@Modifying
 	@Transactional
 	@Query(value = "insert into vozilo_rezervacije (rezervacije_id, vozilo_id) VALUES (?1, ?2)", nativeQuery = true)
-	void insertRezervacijaVozila(Long rezervacija_id, Long vozilo_id);
+	void insertRezervacijaVozila(Integer rezervacija_id, Integer vozilo_id);
 }
