+++ ISA_AMA_projekat/src/main/java/com/example/ISA_AMA_projekat/repository/VoziloRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ISA_AMA_projekat.model.Vozilo;
 
-public interface VoziloRepository extends JpaRepository<Vozilo, Long>{
+public interface VoziloRepository extends JpaRepository<Vozilo, Integer>{
 
 }
