+++ ISA_AMA_projekat/src/main/java/com/example/ISA_AMA_projekat/service/RentacarService.java
@@ -23,7 +23,7 @@ public class RentacarService {
 		return rentRepository.findAll();
 	}
 	
-	public Optional<RentacarServis> findById(Long id){
+	public Optional<RentacarServis> findById(Integer id){
 		return rentRepository.findById(id);
 	}
